Optmize the way we fetch events. (see details)
The way the old way was implemented is that at each iteration we used to
fetch the last block number, calculate the destination block number and
then fetch the events. And as long as we did not reach the latest block
(which is the case most of the time) we would repeat the process without
a cooldown. This was a waste of resources.

The new implementation is that we fetch the last block number and never
update it, untill we fully catch up with the latest block. This way we
can fetch the events in a fewer calls, without even cooling down, until
we reach the latest block. That's when we know we are fully synced and
until then we only update the latest block.

From now on, fetching the latest block should not be in a hot loop.
as we only fetch it after `polling interval` milliseconds.